src/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
executing==2.2.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
logfire==3.9.0
markdown-it-py==3.0.0
mdurl==0.1.2
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
packaging==24.2
protobuf==5.29.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
SQLAlchemy==2.0.39
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
